# Revive model picking refactoring